requirements.txt
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
antspyx==0.5.4
certifi==2025.1.31
charset-normalizer==3.4.1
colorlog==6.9.0
contourpy==1.1.1
cycler==0.12.1
eval_type_backport==0.2.2
filelock==3.16.1
fonttools==4.56.0
fsspec==2025.3.0
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.7.5
monai==1.3.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.0.3
patsy==1.0.1
pillow==10.4.0
pip==24.2
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.21.0
scipy==1.10.1
setuptools==75.1.0
SimpleITK==2.4.1
six==1.17.0
statsmodels==0.14.1
sympy==1.13.3
tifffile==2023.7.10
torch==2.4.1
tqdm==4.67.1
triton==3.0.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.2.3
webcolors==24.8.0
wheel==0.44.0
zipp==3.20.2
